requirements.txt
mcp[cli]
google-auth
google-auth-oauthlib
google-api-python-client